chore: moved content into `src`
