+++ .env.example
@@ -1,2 +1,2 @@
 AMAZON_USER_POOL_ID=YOUR-AMAZON-USERPOOL-ID
-AMAZON_CLIENT_ID=the app client id gotten from "General Settings > "App clients" value.
+AMAZON_CLIENT_ID=the app client id from "General Settings > "App clients" value.
