+++ README.md
@@ -1,4 +1,4 @@
-# Adding SIM Swap Detection to your React Native AWS Apps with Amazon Cognito & **tru.ID** PhoneCheck API
+# SIM Card Based Mobile Authentication with React Native Amazon Cognito & **tru.ID** PhoneCheck API
 
 ## Requirements
 
@@ -12,13 +12,19 @@
 Clone the starter-files branch via:
 
 ```bash
-git clone -b starter-files --single-branch https://github.com/tru-ID/amazon-cognito-sim-swap-detection.git
+git clone -b starter-files --single-branch https://github.com/tru-ID/amazon-cognito-sim-authentication.git
 ```
 
 If you're only interested in the finished code in main then run:
 
 ```bash
-git clone -b main https://github.com/tru-ID/amazon-cognito-sim-swap-detection.git
+git clone -b main https://github.com/tru-ID/amazon-cognito-sim-authentication.git
+```
+
+Install dependencies:
+
+```bash
+npm install
 ```
 
 Create a [tru.ID Account](https://tru.id)
@@ -37,13 +43,13 @@ Install the **tru.ID** CLI [development server plugin](https://github.com/tru-ID
 Create a new **tru.ID** project within the root directory via:
 
 ```
-tru projects:create rn-amazon-auth
+tru projects:create rn-amazon-auth --project-dir .
 ```
 
-Run the development server, pointing it to the directly containing the newly created project configuration. This will also open up a localtunnel to your development server making it publicly accessible to the Internet so that your mobile phone can access it when only connected to mobile data.
+Run the development server, pointing it to the directory containing the newly created project configuration. This will also open up a localtunnel to your development server making it publicly accessible to the Internet so that your mobile phone can access it when only connected to mobile data.
 
 ```
-tru server -t --project-dir ./rn-amazon-auth
+tru server -t
 ```
 
 ## Setting up Amazon Cognito
@@ -53,28 +59,30 @@ Create or log into an [AWS account](https://console.aws.amazon.com)
 Navigate to the AWS Management portal
 
 Search for cognito, and select the service as shown below.
-![Cognito Service](./images/Service.PNG)
+
+![Cognito Service](./src/images/Service.PNG)
 
 You'll be taken to the following page, select "manage user pools" & create a new user pool.
-![Cognito Home](./images/cognito-home.PNG)
+
+![Cognito Home](./src/images/cognito-home.PNG)
 
 Give the pool a name and select "step through settings" as shown below.
 
-![Cognito Pool Name](./images/pool-name.PNG)
+![Cognito Pool Name](./src/images/pool-name.PNG)
 
 Under "attributes" select "Email address or phone number" as shown below.
 
-![Cognito Home](./images/attributes-1.PNG)
+![Cognito Home](./src/images/attributes-1.PNG)
 
 Under the "Which standard attributes do you want to require?" heading select "phone number" click "Next Step" to continue.
 
 Under "policies" keep the defaults, as shown below.
 
-![Cognito Policies](./images/policies.PNG)
+![Cognito Policies](./src/images/policies.PNG)
 
-Under "MFA and verifications" keep the defaults except under the "Which attributes do you want to verify?" heading. There, select "no verification" as shown below.
+Under "MFA and verifications" keep the defaults except under the "Which attributes do you want to verify?" heading. There, select "no verification" as shown below. When that is selected, a banner would be displayed informing you that you haven't selected any verification. It is safe to ignore.
 
-![Cognito Home](./images/MFA.PNG)
+![Cognito Home](./src/images/MFA.PNG)
 
 Under "Message customizations" keep the defaults and proceed to the next step.
 
@@ -83,19 +91,19 @@ Under "Tags" simply proceed to the next step.
 Under "devices" keep the default and proceed to the next step.
 
 Under "App clients" click "Add an app client". Give the App client name a name and ensure to uncheck the "Generate client secret" checkbox because the AWS SDK [doesn't support apps that have a client secret](https://github.com/aws-amplify/amplify-js/tree/master/packages/amazon-cognito-identity-js#configuration) as shown below.
-![Cognito App client name](./images/App-Client-Name.PNG)
+![Cognito App client name](./src/images/App-Client-Name.PNG)
 
 Below, under Auth Flows Configuration ensure to check the "Enable username password based authentication (ALLOW_USER_PASSWORD_AUTH)"
 
 You will then be informed that you have added an app client, as shown below, click "Next Step" to proceed.
 
-![Cognito App client name](./images/App-Client-Name-2.PNG)
+![Cognito App client name](./src/images/App-Client-Name-2.PNG)
 
 Under "Triggers" leave the defaults and proceed to the next step
 
 Finally, under "Reviews" you will be shown a recap of all your settings and allowed to modify them and create the pool. Once you've looked over your settings create the pool. An example is shown below.
 
-![Cognito App client name](./images/App-Review.PNG)
+![Cognito App client name](./src/images/Review.PNG)
 
 You should then be given a `Pool id` & `Pool ARN`. Open the terminal and run the following
 
@@ -103,14 +111,14 @@ You should then be given a `Pool id` & `Pool ARN`. Open the terminal and run the
 cp .env.example .env
 ```
 
-Replace `AMAZON_USER_POOL_ID` with the `Pool id` value and `AMAZON_CLIENT_ID` with the app client id gotten from "General Settings > "App clients" value.
+Replace `AMAZON_USER_POOL_ID` with the `Pool id` value and `AMAZON_CLIENT_ID` with the `App client id` from "General Settings > "App clients" value.
 
-To start the project, ensure you have a physical device connected (see Running React Native on a physical device guide ) then run:
+To start the project, ensure you have a physical device connected (see [Running React Native on a physical device guide](https://reactnative.dev/docs/running-on-device) ) then run:
 
 ```bash
-npm run android
+ npm run android
 #or
-npm run ios
+ npm run ios
 ```
 
 ## References
